#151 Split up RenderendComponent
