+++ apps/lundin/src/app/modules/renderend/renderend.component.html
@@ -9,8 +9,6 @@
 	</div>
 </div>
 
-<div class="game-area">
-	<canvas #canvas
-	(click)="clickCanvas($event)">
-	</canvas>
+<div #gameArea class="game-area">
+	<lundin-renderend-game></lundin-renderend-game>
 </div>
